refactor(torznab): split Torznab API into focused modules
Break the oversized torznab API module into smaller units with clearer responsibility boundaries and keep api.py as the route orchestration layer.

Extract shared helpers into a dedicated module for language defaults, integer coercion, and ordered de-duplication.

Move SkyHook identifier/title resolution and metadata-based season episode discovery into a separate module with isolated cache logic.


Move tvsearch-specific discovery and item emission logic into its own module, including fast season mode language discovery and mapped special probing.

Extract generic preview search and special-search handling into separate search handler functions while preserving existing behavior and test monkeypatch seams.

Reduce torznab/api.py from 1834 lines to 609 lines without changing the Torznab endpoint contract.
